[IS-1086] refactoring grpc-related
 - remove grpc Request in message_code
 - move message_code to utils
 - remove protos package
