+++ setup.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 import os
-import re
 
 from setuptools import setup, find_packages
