+++ tests/dispatcher/test_ws_dispatcher.py
@@ -5,7 +5,7 @@ from mock import AsyncMock
 from jsonrpcclient.requests import Request
 
 from iconrpcserver.dispatcher.default.websocket import WSDispatcher
-from iconrpcserver.protos import message_code
+from iconrpcserver.utils import message_code
 from iconrpcserver.utils.message_queue.stub_collection import StubCollection
 from .conftest import CHANNEL_STUB_NAME
